+++ src/mod/checkStatus.js
