+++ src/modules/initial.js
@@ -1,0 +1,19 @@
+let listTasks = [
+  {
+    description: "Wash the dishes",
+    status: false,
+    index: 0,
+  },
+  {
+    description: "Brush your teeth",
+    status: true,
+    index: 1,
+  },
+  {
+    description: "Wake up early",
+    status: false,
+    index: 2,
+  },
+];
+
+export { listTasks };
