+++ src/modules/store.js
@@ -1,0 +1,9 @@
+import { listTasks } from './modules/initial.js';
+
+export const getDataLocalStorage = () => {
+    return localStorage.getItem('TaskData') || listTasks;
+}
+
+export const setDataLocalStorage = (listTasks) => {
+    localStorage.setItem('TaskData', JSON.stringify(listTasks));
+}
